Blackjack finalized, cash converted to double

+++ src/main/java/io/intellijokers/Game.java
@@ -25,8 +25,7 @@ public class Game {
         return wager;
     }
 
-    public double payout(int odds, User player) {
-        return 0;
+    public double payout(double odds, User player) { return 0;
     }
 
     public double payout(User player) {
